% Auto-generated: do not edit by hand
\docType{package}
\name{dashPivottable-package}
\alias{dashPivottable}
\title{Interactive React-based Pivot Tables for Dash }
\description{
Pivot tables are useful for interactive presentation of  summary statistics computed for data contained in another table. The 'dashPivottable' package wraps 'react-pivottable', making it easy to add drag-and-drop tables into your Dash for R applications.
}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plotly.com>
}
